Removes $(window).scroll considerations
Since styling of container is set to use fixed positioning, taking the window scroll distances into account could cause the age-check container to be unaccessible by the user. An unaccessible dialog can happen if a page's content is long in content and the user managed to scroll before the $(document).ready events fire.